% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_random_rows.R
\name{assign_random_rows}
\alias{assign_random_rows}
\title{\code{Assign genotypes to random rows.}}
\usage{
assign_random_rows(rows = c(1001, 1002), genos = c("Genotype1",
  "Genotype2"))
}
\arguments{
\item{rows}{A vector of row numbers. [vector, =c(1001, 1002)]}

\item{genos}{A vector of genotypes with the same length with rows. [vector, =c("G1", "G2")]}
}
\value{
return a data.frame with rows and genos
}
\description{
\code{Assign genotypes to random rows.}
}
\examples{
assign_random_rows()

}
